Add setSquare function and linesForSquare array
The setSqare function simply sets a value of a square to a given value.
The linesForSquare array contains nine arrays of strings, one array of
strings for each position in the tray. The strings are keys for the lines
object, which gives coordinates for possible winning lines.

The plan is to use this array to efficiently and readably call the
winCheck function on every possible line on a given square after a player
plays that square.